@@ -1,44 +1,44 @@
 {
    "name": "output",
-   "version": "0.2.6-beta",
+   "version": "1.0.0-beta",
    "description": "Publication management for universities.",
    "devDependencies": {
-      "@types/cookie-parser": "^1.4.5",
-      "@types/multer": "^1.4.10",
-      "@types/node": "^20.9.0",
-      "ts-node": "^10.9.1",
-      "typescript": "^5.2.2"
+      "@types/cookie-parser": "^1.4.7",
+      "@types/multer": "^1.4.12",
+      "@types/node": "^22.7.5",
+      "ts-node": "^10.9.2",
+      "typescript": "^5.6.3"
    },
    "dependencies": {
-      "@nestjs/axios": "3.0.1",
-      "@nestjs/cli": "^10.2.1",
-      "@nestjs/common": "^10.2.8",
-      "@nestjs/config": "^3.1.1",
-      "@nestjs/core": "^10.2.8",
+      "@nestjs/axios": "3.0.3",
+      "@nestjs/cli": "^10.4.5",
+      "@nestjs/common": "^10.4.4",
+      "@nestjs/config": "^3.2.3",
+      "@nestjs/core": "^10.4.4",
       "@nestjs/jwt": "^10.2.0",
-      "@nestjs/passport": "^10.0.2",
-      "@nestjs/platform-express": "^10.2.8",
-      "@nestjs/swagger": "^7.1.15",
-      "@nestjs/typeorm": "^10.0.0",
-      "axios": "^1.6.1",
-      "body-parser": "^1.20.2",
+      "@nestjs/passport": "^10.0.3",
+      "@nestjs/platform-express": "^10.4.4",
+      "@nestjs/swagger": "^7.4.2",
+      "@nestjs/typeorm": "^10.0.2",
+      "axios": "^1.7.7",
+      "body-parser": "^1.20.3",
       "class-transformer": "^0.5.1",
-      "class-validator": "^0.14.0",
-      "cookie-parser": "^1.4.6",
+      "class-validator": "^0.14.1",
+      "cookie-parser": "^1.4.7",
       "cross-env": "^7.0.3",
-      "express": "^4.18.2",
+      "express": "^4.21.1",
       "https": "^1.0.0",
-      "moment": "^2.29.4",
+      "moment": "^2.30.1",
       "mysql": "^2.18.1",
-      "node": "^20.9.0",
+      "node": "^20.18.0",
       "papaparse": "^5.4.1",
-      "passport": "^0.6.0",
+      "passport": "^0.7.0",
       "passport-local": "^1.0.0",
-      "pg": "^8.11.3",
-      "reflect-metadata": "^0.1.13",
+      "pg": "^8.13.0",
+      "reflect-metadata": "^0.2.2",
       "rxjs": "^7.8.1",
-      "swagger-ui-express": "^5.0.0",
-      "typeorm": "^0.3.17",
+      "swagger-ui-express": "^5.0.1",
+      "typeorm": "^0.3.20",
       "xml-js": "^1.6.11"
    },
    "scripts": {
