@@ -1,4 +1,4 @@
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MAT_MOMENT_DATE_ADAPTER_OPTIONS, MatMomentDateModule } from '@angular/material-moment-adapter';
@@ -75,99 +75,98 @@ import { CombineDialogComponent } from './tools/combine-dialog/combine-dialog.co
 import { ConfirmDialogComponent } from './tools/confirm-dialog/confirm-dialog.component';
 import { CsvFormatComponent } from './tools/csv-format/csv-format.component';
 import { LogDialogComponent } from './tools/log-dialog/log-dialog.component';
-import { SelectInstituteDialogComponent } from './tools/select-institute-dialog/select-institute-dialog.component';
 import { TableComponent } from './tools/table/table.component';
 import { DeletePublicationDialogComponent } from './tools/delete-publication-dialog/delete-publication-dialog.component';
 import { AliasFormComponent } from './tools/alias-form/alias-form.component';
+import { AuthorshipFormComponent } from './pages/windows/authorship-form/authorship-form.component';
+import { RolesComponent } from './pages/master-data/roles/roles.component';
+import { RoleFormComponent } from './pages/windows/role-form/role-form.component';
 
 export const metaReducers: MetaReducer[] = [hydrationMetaReducer];
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    StartComponent,
-    PublicationsComponent,
-    TableComponent,
-    BreadcrumpComponent,
-    PublicationFormComponent,
-    ConfirmDialogComponent,
-    ReportingYearFormComponent,
-    CombineDialogComponent,
-    FilterViewComponent,
-    GreaterEntityFormComponent,
-    PublisherFormComponent,
-    ContractFormComponent,
-    SelectInstituteDialogComponent,
-    AuthorsComponent,
-    AuthorFormComponent,
-    InstitutionsComponent,
-    InstituteFormComponent,
-    ContractsComponent,
-    PublishersComponent,
-    GreaterEntitiesComponent,
-    FundersComponent,
-    FunderFormComponent,
-    PubTypesComponent,
-    PubTypeFormComponent,
-    OaCategoriesComponent,
-    OaCategoryFormComponent,
-    PlausibilityComponent,
-    ImportComponent,
-    EnrichComponent,
-    ExportComponent,
-    StatisticsComponent,
-    LogDialogComponent,
-    ImportConfigComponent,
-    CsvFormatComponent,
-    StatisticsYearComponent,
-    InvoiceFormComponent,
-    CostItemFormComponent,
-    CostCenterComponent,
-    CostTypesComponent,
-    CostCenterFormComponent,
-    CostTypeFormComponent,
-    DeletePublicationDialogComponent,
-    AliasFormComponent
-  ],
-  imports: [
-    BrowserModule,
-    MatTableModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    MatIconModule,
-    MatMenuModule,
-    MatToolbarModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    MatPaginatorModule,
-    MatSortModule,
-    MatFormFieldModule,
-    FormsModule,
-    MatSelectModule,
-    MatInputModule,
-    MatDatepickerModule,
-    MatMomentDateModule,
-    ReactiveFormsModule,
-    MatDialogModule,
-    MatTooltipModule,
-    MatExpansionModule,
-    MatSnackBarModule,
-    MatAutocompleteModule,
-    MatChipsModule,
-    MatProgressBarModule,
-    MatDividerModule,
-    MatCardModule,
-    HighchartsChartModule,
-    NgbModule,
-    HttpClientModule,
-    StoreModule.forRoot({ viewConfigReducer }, { metaReducers }),
-    !environment.production ? StoreDevtoolsModule.instrument() : []
-  ],
-  providers: [
-    { provide: MatPaginatorIntl, useClass: MatPaginatorDE },
-    { provide: MAT_DATE_LOCALE, useValue: 'de-DE' },
-    { provide: MAT_MOMENT_DATE_ADAPTER_OPTIONS, useValue: { useUtc: true } },
-    { provide: AuthorizationService, useClass: environment.authorization_service}],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        StartComponent,
+        PublicationsComponent,
+        TableComponent,
+        BreadcrumpComponent,
+        PublicationFormComponent,
+        ConfirmDialogComponent,
+        ReportingYearFormComponent,
+        CombineDialogComponent,
+        FilterViewComponent,
+        GreaterEntityFormComponent,
+        PublisherFormComponent,
+        ContractFormComponent,
+        AuthorsComponent,
+        AuthorFormComponent,
+        InstitutionsComponent,
+        InstituteFormComponent,
+        ContractsComponent,
+        PublishersComponent,
+        GreaterEntitiesComponent,
+        FundersComponent,
+        FunderFormComponent,
+        PubTypesComponent,
+        PubTypeFormComponent,
+        OaCategoriesComponent,
+        OaCategoryFormComponent,
+        PlausibilityComponent,
+        ImportComponent,
+        EnrichComponent,
+        ExportComponent,
+        StatisticsComponent,
+        LogDialogComponent,
+        ImportConfigComponent,
+        CsvFormatComponent,
+        StatisticsYearComponent,
+        InvoiceFormComponent,
+        CostItemFormComponent,
+        CostCenterComponent,
+        CostTypesComponent,
+        CostCenterFormComponent,
+        CostTypeFormComponent,
+        DeletePublicationDialogComponent,
+        AliasFormComponent,
+        AuthorshipFormComponent,
+        RolesComponent,
+        RoleFormComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        MatTableModule,
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        MatIconModule,
+        MatMenuModule,
+        MatToolbarModule,
+        MatButtonModule,
+        MatCheckboxModule,
+        MatPaginatorModule,
+        MatSortModule,
+        MatFormFieldModule,
+        FormsModule,
+        MatSelectModule,
+        MatInputModule,
+        MatDatepickerModule,
+        MatMomentDateModule,
+        ReactiveFormsModule,
+        MatDialogModule,
+        MatTooltipModule,
+        MatExpansionModule,
+        MatSnackBarModule,
+        MatAutocompleteModule,
+        MatChipsModule,
+        MatProgressBarModule,
+        MatDividerModule,
+        MatCardModule,
+        HighchartsChartModule,
+        NgbModule,
+        StoreModule.forRoot({ viewConfigReducer }, { metaReducers }),
+        !environment.production ? StoreDevtoolsModule.instrument({connectInZone: true}) : []], providers: [
+        { provide: MatPaginatorIntl, useClass: MatPaginatorDE },
+        { provide: MAT_DATE_LOCALE, useValue: 'de-DE' },
+        { provide: MAT_MOMENT_DATE_ADAPTER_OPTIONS, useValue: { useUtc: true } },
+        { provide: AuthorizationService, useClass: environment.authorization_service },
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 export class AppModule { }
