@@ -26,6 +26,10 @@ import { AliasFunder } from "./entity/alias/AliasFunder";
 import { Language } from "./entity/Language";
 import config from "../config";
 import { PublisherDOI } from "./entity/PublisherDOI";
+import { Role } from "./entity/Role";
+import { PublicationIdentifier } from "./entity/PublicationIdentifier";
+import { AliasAuthorLastName } from "./entity/alias/AliasAuthorLastName";
+import { AliasAuthorFirstName } from "./entity/alias/AliasAuthorFirstName";
 
 //export const init_service = require(config().init_service)
 //export const init_service = import (config().init_service);
@@ -42,8 +46,8 @@ import { PublisherDOI } from "./entity/PublisherDOI";
         inject: [DatabaseConfigService],
       }),
       TypeOrmModule.forFeature([Author, AuthorPublication, Contract, CostCenter, CostItem, CostType, Funder, GreaterEntity, Identifier, 
-        Institute, Invoice, OA_Category, Publication, PublicationType, Publisher, PublisherDOI, Config, Language,
-        AliasInstitute, AliasPublisher, AliasPubType, AliasFunder])
+        Institute, Invoice, OA_Category, Publication, PublicationType, Publisher, PublisherDOI, Config, Language, Role, PublicationIdentifier, 
+        AliasInstitute, AliasPublisher, AliasPubType, AliasFunder, AliasAuthorFirstName, AliasAuthorLastName])
     ],
     providers: [
       DatabaseConfigService, 
