@@ -24,11 +24,11 @@ async function bootstrap() {
     const cert_passphrase: string = app.get(ConfigService).get<string>('APP_SSL_PASSPHRASE');
     const cors_origins: string[] = app.get(ConfigService).get<string>('APP_CORS_ORIGINS').split(',');
     const base_path: string = app.get(ConfigService).get<string>('APP_BASE_PATH');
-
+    
     const config = new DocumentBuilder()
         .setTitle('Output API')
         .setDescription('The Output API description')
-        .setVersion('0.2.6-beta')
+        .setVersion('1.0.0-beta')
         .addServer(base_path)
         .build();
     const options: SwaggerDocumentOptions = {
