@@ -13,7 +13,7 @@ export class ContractService {
     constructor(@InjectRepository(Contract) private repository: Repository<Contract>, private configService:ConfigService, private publicationService:PublicationService) { }
 
     public get() {
-        return this.repository.find();
+        return this.repository.find({relations:{publisher:true}});
     }
     
     public async one(id:number, writer:boolean):Promise<Contract> {
@@ -113,7 +113,7 @@ export class ContractService {
     
     public async delete(insts:Contract[]) {
         for (let inst of insts) {
-            let conE: Contract = await this.repository.findOne({where:{id:inst.id},relations:{publisher:true, publications:true}});
+            let conE: Contract = await this.repository.findOne({where:{id:inst.id},relations:{publisher:true, publications:true},withDeleted: true});
             let pubs = [];
             if (conE.publications) for (let pub of conE.publications) {
                 pubs.push({id:pub.id, contract: null})
