@@ -52,11 +52,8 @@ export class PublisherService {
         if (!publisher_ent && publisher.doi_prefixes) {
             publisher_ent = await this.repository.findOne({ where: { doi_prefixes: { doi_prefix: In(publisher.doi_prefixes.map(e => e.doi_prefix)) } }, relations: { doi_prefixes: true } })
         }
-        if (publisher_ent) {
-            if (!publisher_ent.location && publisher.location) publisher_ent = await this.repository.save({ id:publisher_ent.id, location: publisher.location });
-            return publisher_ent;
-        }
-        else return this.repository.save({ label, location:publisher.location, doi_prefixes:publisher.doi_prefixes });
+        if (publisher_ent) return publisher_ent;
+        else return this.repository.save({ label, doi_prefixes:publisher.doi_prefixes });
     }
 
     public async identifyPublisher(title: string) {
@@ -88,19 +85,15 @@ export class PublisherService {
                 .select("publisher.id", "id")
                 .addSelect("publisher.label", "label")
                 .addSelect("STRING_AGG(DISTINCT doi_prefix.doi_prefix, ';')", "doi_prefix")
-                .addSelect("publisher.location", "location")
                 .groupBy("publisher.id")
-                .addGroupBy("publisher.location")
                 .addGroupBy("publisher.label")
             , "a")
             .leftJoin(Publication, "publication", "publication.\"publisherId\" = a.id and publication.pub_date between :beginDate and :endDate", { beginDate, endDate })
             .select("a.id", "id")
             .addSelect("a.label", "label")
-            .addSelect("a.doi_prefix", "doi_prefix")
-            .addSelect("a.location", "location")            
+            .addSelect("a.doi_prefix", "doi_prefix")      
             .addSelect("COUNT(publication.id)", "pub_count")
             .groupBy("a.id")
-            .addGroupBy("a.location")
             .addGroupBy("a.label")
             .addGroupBy("a.doi_prefix")
 
@@ -127,7 +120,6 @@ export class PublisherService {
             }
             await this.publicationService.save(pubs)
             if (!res.label && aut.label) res.label = aut.label;
-            if (!res.location && aut.location) res.location = aut.location;
             if (!res.aliases) res.aliases = [];
             res.aliases = res.aliases.concat(aut.aliases.map(e => { return { alias: e.alias, elementId: aut1.id } }))
         }
@@ -148,7 +140,7 @@ export class PublisherService {
 
     public async delete(insts: Publisher[]) {
         for (let inst of insts) {
-            let conE: Publisher = await this.repository.findOne({ where: { id: inst.id }, relations: { publications: true, aliases: true } });
+            let conE: Publisher = await this.repository.findOne({ where: { id: inst.id }, relations: { publications: true, aliases: true }, withDeleted: true});
             let pubs = [];
             if (conE.publications) for (let pub of conE.publications) {
                 pubs.push({ id: pub.id, publisher: null })
