@@ -22,6 +22,7 @@ import { InstitutionService } from '../entities/institution.service';
 import { LanguageService } from '../entities/language.service';
 import { Publisher } from '../../entity/Publisher';
 import { GreaterEntity } from '../../entity/GreaterEntity';
+import { RoleService } from '../entities/role.service';
 
 @Injectable()
 export class BASEImportService extends ApiImportOffsetService {
@@ -29,9 +30,9 @@ export class BASEImportService extends ApiImportOffsetService {
     constructor(protected publicationService: PublicationService, protected authorService: AuthorService,
         protected geService: GreaterEntityService, protected funderService: FunderService, protected publicationTypeService: PublicationTypeService,
         protected publisherService: PublisherService, protected oaService: OACategoryService, protected contractService: ContractService,
-        protected costTypeService: CostTypeService, protected reportService: ReportItemService, protected instService: InstitutionService, protected languageService: LanguageService, protected configService: ConfigService,
+        protected costTypeService: CostTypeService, protected reportService: ReportItemService, protected instService: InstitutionService, protected languageService: LanguageService, protected roleService: RoleService, protected configService: ConfigService,
         protected http: HttpService) {
-        super(publicationService, authorService, geService, funderService, publicationTypeService, publisherService, oaService, contractService, costTypeService, reportService, instService, languageService, configService, http);
+        super(publicationService, authorService, geService, funderService, publicationTypeService, publisherService, oaService, contractService, costTypeService, reportService, instService, languageService, roleService, configService, http);
         this.configService.get('searchTags').forEach(tag => {
             this.searchText += tag + "+OR+"
         })
@@ -56,11 +57,10 @@ export class BASEImportService extends ApiImportOffsetService {
         license: UpdateOptions.REPLACE_IF_EMPTY,
         invoice: UpdateOptions.IGNORE,
         status: UpdateOptions.IGNORE,
-        editors :UpdateOptions.IGNORE,
-        abstract :UpdateOptions.IGNORE,
-        citation :UpdateOptions.IGNORE,
-        page_count :UpdateOptions.IGNORE,
-        peer_reviewed :UpdateOptions.IGNORE,
+        abstract: UpdateOptions.IGNORE,
+        citation: UpdateOptions.IGNORE,
+        page_count: UpdateOptions.IGNORE,
+        peer_reviewed: UpdateOptions.IGNORE,
     };
     //e.g.https://api.base-search.net/cgi-bin/BaseHttpSearchInterface.fcgi?func=PerformSearch&format=json&query=(guericke)
     protected url = 'https://api.base-search.net/cgi-bin/BaseHttpSearchInterface.fcgi?';
@@ -188,13 +188,10 @@ export class BASEImportService extends ApiImportOffsetService {
     protected getStatus(element: any): number {
         return 0;
     }
-    protected getEditors(element: any): string {
-        return null;
-    }
     protected getAbstract(element: any): string {
         return null;
     }
-    protected getCitation(element: any): string {
+    protected getCitation(element: any): {volume?:string, issue?: string, first_page?: string, last_page?: string, publisher_location?: string, edition?: string, article_number?: string} {
         return null;
     }
     protected getPageCount(element: any): number {
