@@ -1,6 +1,6 @@
 import { HttpService } from '@nestjs/axios';
 import { ConflictException, Injectable } from '@nestjs/common';
-import { catchError, EMPTY, mergeAll, Observable, of, queueScheduler, scheduled } from 'rxjs';
+import { catchError, delay, EMPTY, mergeAll, Observable, of, queueScheduler, scheduled } from 'rxjs';
 import { FindManyOptions } from 'typeorm';
 import { AuthorService } from '../entities/author.service';
 import { ContractService } from '../entities/contract.service';
@@ -21,6 +21,7 @@ import { GreaterEntity } from '../../entity/GreaterEntity';
 import { Invoice } from '../../entity/Invoice';
 import { Publisher } from '../../entity/Publisher';
 import { UpdateMapping, UpdateOptions } from '../../../../output-interfaces/Config';
+import { RoleService } from '../entities/role.service';
 
 @Injectable()
 /**
@@ -45,7 +46,6 @@ export class DOAJEnrichService extends AbstractImportService {
         license: UpdateOptions.IGNORE,
         invoice: UpdateOptions.IGNORE,
         status: UpdateOptions.IGNORE,
-        editors: UpdateOptions.IGNORE,
         abstract: UpdateOptions.IGNORE,
         citation: UpdateOptions.IGNORE,
         page_count: UpdateOptions.IGNORE,
@@ -77,8 +77,6 @@ export class DOAJEnrichService extends AbstractImportService {
                 identifiers: [{type:'issn', value: element['query'].split(':')[1]}]
             }
         } else {
-            console.log(element['results'][0]['bibjson']['oa_start'])
-            console.log(new Date(element['results'][0]['bibjson']['oa_start'],0))
             return {
                 label: element['results'][0]['bibjson']['title'],
                 identifiers: [{type:'issn', value: element['results'][0]['bibjson']['eissn']}],
@@ -119,13 +117,10 @@ export class DOAJEnrichService extends AbstractImportService {
     protected getStatus(element: any): number {
         return null;
     }
-    protected getEditors(element: any): string {
-        return null;
-    }
     protected getAbstract(element: any): string {
         return null;
     }
-    protected getCitation(element: any): string {
+    protected getCitation(element: any): {volume?:string, issue?: string, first_page?: string, last_page?: string, publisher_location?: string, edition?: string, article_number?: string} {
         return null;
     }
     protected getPageCount(element: any): number {
@@ -138,8 +133,8 @@ export class DOAJEnrichService extends AbstractImportService {
     constructor(protected publicationService: PublicationService, protected authorService: AuthorService,
         protected geService: GreaterEntityService, protected funderService: FunderService, protected publicationTypeService: PublicationTypeService,
         protected publisherService: PublisherService, protected oaService: OACategoryService, protected contractService: ContractService,
-        protected costTypeService: CostTypeService, protected reportService: ReportItemService, protected instService: InstitutionService, protected languageService: LanguageService, protected configService: ConfigService, protected http: HttpService) {
-        super(publicationService, authorService, geService, funderService, publicationTypeService, publisherService, oaService, contractService, costTypeService, reportService, instService, languageService, configService);
+        protected costTypeService: CostTypeService, protected reportService: ReportItemService, protected instService: InstitutionService, protected languageService: LanguageService,  protected roleService: RoleService, protected configService: ConfigService, protected http: HttpService) {
+        super(publicationService, authorService, geService, funderService, publicationTypeService, publisherService, oaService, contractService, costTypeService, reportService, instService, languageService, roleService, configService);
     }
 
     private publicationsUpdate = [];
@@ -173,7 +168,7 @@ export class DOAJEnrichService extends AbstractImportService {
 
     private request(issn: string): Observable<any> {
         let url = this.createUrl(issn);
-        return this.http.get(url).pipe(catchError((error, caught) => {
+        return this.http.get(url).pipe(delay(100),catchError((error, caught) => {
             if (error.response?.status === 404) {
                 this.reportService.write(this.report, { type: 'warning', timestamp: new Date(), origin: 'import', text: 'Not found: ' + issn })
             } else if (error.response?.status === 422) {
@@ -221,21 +216,21 @@ export class DOAJEnrichService extends AbstractImportService {
         }
 
         for (let pub of this.uniqueGE) obs$.push(this.request(pub.greater_entity.identifiers.find(e => e.type === 'issn').value));
-        //console.log('Started enrich ' + this.name + ' for ' + publications.length + ' publications');
         this.reportService.write(this.report, { type: 'info', timestamp: new Date(), origin: this.name, text: `Starting import with where clause ${this.whereClause.toString()}` })
         this.reportService.write(this.report, { type: 'info', timestamp: new Date(), origin: this.name, text: `${publications.length} elements found` })
         scheduled(obs$, queueScheduler).pipe(mergeAll(this.parallelCalls)).subscribe({
             next: async (data: any) => {
                 if (data) {
                     let item = this.getData(data);
-                    console.log(item['query'].split(':')[1])
+                    //console.log(item['query'].split(':')[1])
                     if (item) {
                         let orig = this.uniqueGE.find(e => {
-                            if (!e.greater_entity.identifiers) {
-                                console.log(e.id)
-                            }
-                            return e.greater_entity.identifiers.find(e => e.type === 'issn').value === item['query'].split(':')[1]
+                            if (e.greater_entity.identifiers) return e.greater_entity.identifiers.find(e => e.type === 'issn').value === item['query'].split(':')[1]
                         })
+                        if (!orig) {
+                            this.reportService.write(this.report, { type: 'error', publication_id: orig?.id, timestamp: new Date(), origin: 'mapUpdate', text: 'no publication to update could be found' })
+                            return null;
+                        }
                         let pubUpd = await this.mapUpdate(item, orig).catch(e => {
                             this.reportService.write(this.report, { type: 'error', publication_id: orig?.id, timestamp: new Date(), origin: 'mapUpdate', text: e.stack ? e.stack : e.message })
                             //console.log('Error while mapping update for publication ' + orig.id + ': ' + e.message)
@@ -250,7 +245,7 @@ export class DOAJEnrichService extends AbstractImportService {
                 } //else this.errors++;
 
                 // Update Progress Value
-                if (this.progress !== 0) this.progress = (this.processedPublications + this.errors) / publications.length;
+                if (this.progress !== 0) this.progress = (this.processedPublications + this.errors) / this.uniqueGE.length;
                 if (this.progress === 1) {
                     //console.log(this.publicationsUpdate.length + ' pubs update to DB, ' + this.errors + ' errors');
                     //finalize
