@@ -1,19 +1,21 @@
 export interface FilterOptions {
     corresponding?: boolean,
+    locked?: boolean,
     instituteId?: number,
-    notInstituteId?: number,
+    notInstituteId?: number[],
     publisherId?: number,
-    notPublisherId?: number,
+    notPublisherId?: number[],
     contractId?: number,
-    notContractId?: number,
+    notContractId?: number[],
     pubTypeId?: number,
-    notPubTypeId?: number,
-    oaCatId?: number
-    notOaCatId?: number
+    notPubTypeId?: number[],
+    oaCatId?: number,
+    notOaCatId?: number[]
 }
 
 export interface HighlightOptions {
     corresponding?: boolean,
+    locked?: boolean,
     instituteId?: number,
     publisherId?: number,
     contractId?: number,
