@@ -47,6 +47,12 @@ export class StatisticsYearComponent implements OnInit {
     },
 
   }; // required
+  chartOptionsLocked = {
+    ...this.chartOptions,
+    title: {
+      text: 'Anteil gesperrt'
+    }
+  }
   chartOptionsInstitute = {
     ...this.chartOptions,
     title: {
@@ -70,7 +76,7 @@ export class StatisticsYearComponent implements OnInit {
     title: {
       text: 'Anteil Publikationsarten'
     }
-  } 
+  }
   chartOptionsContract = {
     ...this.chartOptions,
     title: {
@@ -84,17 +90,18 @@ export class StatisticsYearComponent implements OnInit {
   updateFlag3: boolean = false; // set to true if you wish to update the chart
   updateFlag4: boolean = false; // set to true if you wish to update the chart
   updateFlag5: boolean = false; // set to true if you wish to update the chart
+  updateFlagLocked: boolean = false; // set to true if you wish to update the chart
   oneToOneFlag: boolean = true; // changing number of series
   runOutsideAngular: boolean = false; // optional boolean, defaults to false
 
   year;
   costs = false;
 
-  institutes:{id:number, label:string}[] = []
-  oa_cats:{id:number, label:string}[] = []
-  publisher:{id:number, label:string}[] = []
-  constracts:{id:number, label:string}[] = []
-  pub_types:{id:number, label:string}[] = []
+  institutes: { id: number, label: string }[] = []
+  oa_cats: { id: number, label: string }[] = []
+  publisher: { id: number, label: string }[] = []
+  constracts: { id: number, label: string }[] = []
+  pub_types: { id: number, label: string }[] = []
 
   constructor(private route: ActivatedRoute, private statService: StatisticsService, private _snackBar: MatSnackBar) { }
 
@@ -104,9 +111,9 @@ export class StatisticsYearComponent implements OnInit {
     this.loadData(this.costs)
   }
 
-  loadData(costs:boolean) {
+  loadData(costs: boolean) {
     this.costs = costs;
-    let ob$:Observable<any> = this.statService.corresponding(this.year, this.filter).pipe(map(
+    let ob$: Observable<any> = this.statService.corresponding(this.year, this.filter).pipe(map(
       data => {
         let chartData = []
         for (let e of data) {
@@ -117,15 +124,34 @@ export class StatisticsYearComponent implements OnInit {
           type: 'pie',
           name: 'Art',
           data: chartData,
-          events: {click: 
-            (event) => {this.applyFilter(event.point.series.name, event.point.name)}
+          events: {
+            click:
+              (event) => { this.applyFilter(event.point.series.name, event.point.name) }
           }
         }]
         this.updateFlag = true;
       }));
+    ob$ = merge(ob$, this.statService.locked(this.year, this.filter).pipe(map(
+      data => {
+        let chartData = []
+        for (let e of data) {
+          chartData.push(['gesperrt', parseFloat(e.locked)])
+          chartData.push(['nicht gesperrt', parseFloat(e.value) - parseFloat(e.locked)])
+        }
+        this.chartOptionsLocked.series = [{
+          type: 'pie',
+          name: 'Gesperrt',
+          data: chartData,
+          events: {
+            click:
+              (event) => { this.applyFilter(event.point.series.name, event.point.name) }
+          }
+        }]
+        this.updateFlagLocked = true;
+      })));
     ob$ = merge(ob$, this.statService.institute(this.year, costs, this.filter).pipe(map(
       data => {
-        this.institutes = data.map(e => {return {id: e['id'], label: e['institute']}});
+        this.institutes = data.map(e => { return { id: e['id'], label: e['institute'] } });
         let chartData = []
         for (let e of data) {
           chartData.push([e.institute, parseFloat(e.value)])
@@ -134,15 +160,16 @@ export class StatisticsYearComponent implements OnInit {
           type: 'pie',
           name: 'Institut',
           data: chartData,
-          events: {click: 
-            (event) => {this.applyFilter(event.point.series.name, event.point.name)}
+          events: {
+            click:
+              (event) => { this.applyFilter(event.point.series.name, event.point.name) }
           }
         }]
         this.updateFlag1 = true;
       })));
-      ob$ = merge(ob$, this.statService.oaCat(this.year, costs, this.filter).pipe(map(
+    ob$ = merge(ob$, this.statService.oaCat(this.year, costs, this.filter).pipe(map(
       data => {
-        this.oa_cats = data.map(e => {return {id: e['id'], label: e['oa_cat']}});
+        this.oa_cats = data.map(e => { return { id: e['id'], label: e['oa_cat'] } });
         let chartData = []
         for (let e of data) {
           chartData.push([e.oa_cat, parseFloat(e.value)])
@@ -151,15 +178,16 @@ export class StatisticsYearComponent implements OnInit {
           type: 'pie',
           name: 'OA-Kategorie',
           data: chartData,
-          events: {click: 
-            (event) => {this.applyFilter(event.point.series.name, event.point.name)}
+          events: {
+            click:
+              (event) => { this.applyFilter(event.point.series.name, event.point.name) }
           }
         }]
         this.updateFlag2 = true;
       })));
-      ob$ = merge(ob$, this.statService.publisher(this.year, costs, this.filter).pipe(map(
+    ob$ = merge(ob$, this.statService.publisher(this.year, costs, this.filter).pipe(map(
       data => {
-        this.publisher = data.map(e => {return {id: e['id'], label: e['publisher']}});
+        this.publisher = data.map(e => { return { id: e['id'], label: e['publisher'] } });
         let chartData = []
         for (let e of data) {
           chartData.push([e.publisher, parseFloat(e.value)])
@@ -168,15 +196,16 @@ export class StatisticsYearComponent implements OnInit {
           type: 'pie',
           name: 'Verlag',
           data: chartData,
-          events: {click: 
-            (event) => {this.applyFilter(event.point.series.name, event.point.name)}
+          events: {
+            click:
+              (event) => { this.applyFilter(event.point.series.name, event.point.name) }
           }
         }]
         this.updateFlag3 = true;
       })));
-      ob$ = merge(ob$, this.statService.pub_type(this.year, costs, this.filter).pipe(map(
+    ob$ = merge(ob$, this.statService.pub_type(this.year, costs, this.filter).pipe(map(
       data => {
-        this.pub_types = data.map(e => {return {id: e['id'], label: e['pub_type']}});
+        this.pub_types = data.map(e => { return { id: e['id'], label: e['pub_type'] } });
         let chartData = []
         for (let e of data) {
           chartData.push([e.pub_type, parseFloat(e.value)])
@@ -185,15 +214,16 @@ export class StatisticsYearComponent implements OnInit {
           type: 'pie',
           name: 'Publikationsart',
           data: chartData,
-          events: {click: 
-            (event) => {this.applyFilter(event.point.series.name, event.point.name)}
+          events: {
+            click:
+              (event) => { this.applyFilter(event.point.series.name, event.point.name) }
           }
         }]
         this.updateFlag4 = true;
       })));
-      ob$ = merge(ob$, this.statService.contract(this.year, costs, this.filter).pipe(map(
+    ob$ = merge(ob$, this.statService.contract(this.year, costs, this.filter).pipe(map(
       data => {
-        this.constracts = data.map(e => {return {id: e['id'], label: e['contract']}});
+        this.constracts = data.map(e => { return { id: e['id'], label: e['contract'] } });
         let chartData = []
         for (let e of data) {
           chartData.push([e.contract, parseFloat(e.value)])
@@ -202,18 +232,19 @@ export class StatisticsYearComponent implements OnInit {
           type: 'pie',
           name: 'Vertrag',
           data: chartData,
-          events: {click: 
-            (event) => {this.applyFilter(event.point.series.name, event.point.name)}
+          events: {
+            click:
+              (event) => { this.applyFilter(event.point.series.name, event.point.name) }
           }
         }]
         this.updateFlag5 = true;
       })));
-      ob$.subscribe({
-        error: err => this._snackBar.open(`Backend nicht erreichbar`, 'Oh oh!', {
-          panelClass: [`danger-snackbar`],
-          verticalPosition: 'top'
-        })
+    ob$.subscribe({
+      error: err => this._snackBar.open(`Backend nicht erreichbar`, 'Oh oh!', {
+        panelClass: [`danger-snackbar`],
+        verticalPosition: 'top'
       })
+    })
   }
 
   getLink() {
@@ -224,69 +255,122 @@ export class StatisticsYearComponent implements OnInit {
     return '/Berichte/' + this.year;
   }
 
-  filter:FilterOptions;
-  filterText:string=''
+  filter: FilterOptions = {};
+  filterChips: { text: string, key: string, id?: number }[] = [];
+
 
   applyFilter(series_name: string, cat_name: string) {
     if (series_name === 'Art' && cat_name === 'sonstige') return;
-    this.filterText += series_name+': '+cat_name +' ';
+    let key = undefined;
     if (series_name === 'Art') {
-      if (cat_name === 'corresponding') this.filter = {...this.filter, corresponding: true}
-    } 
-    if (series_name === 'Institut') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, instituteId: null}
-      else this.filter = {...this.filter,instituteId: this.institutes.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'OA-Kategorie') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, oaCatId: null}
-      else this.filter = {...this.filter,oaCatId: this.oa_cats.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'Vertrag') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, contractId: null}
-      else this.filter = {...this.filter,contractId: this.constracts.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'Publikationsart') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, pubTypeId: null}
-      else this.filter = {...this.filter,pubTypeId: this.pub_types.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'Verlag') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, publisherId: null}
-      else this.filter = {...this.filter,publisherId: this.publisher.find(e => e.label === cat_name)?.id}
+      if (cat_name === 'corresponding') this.filter = { ...this.filter, corresponding: true }
+      key = 'corresponding'
     }
+    if (series_name === 'Gesperrt') {
+      if (cat_name === 'gesperrt') this.filter = { ...this.filter, locked: true }
+      else this.filter = { ...this.filter, locked: false }
+      key = 'locked'
+    }
+    if (series_name === 'Institut') {
+      if (cat_name === 'Unbekannt') this.filter = { ...this.filter, instituteId: null }
+      else this.filter = { ...this.filter, instituteId: this.institutes.find(e => e.label === cat_name)?.id }
+      key = 'instituteId'
+    }
+    if (series_name === 'OA-Kategorie') {
+      if (cat_name === 'Unbekannt') this.filter = { ...this.filter, oaCatId: null }
+      else this.filter = { ...this.filter, oaCatId: this.oa_cats.find(e => e.label === cat_name)?.id }
+      key = 'oaCatId'
+    }
+    if (series_name === 'Vertrag') {
+      if (cat_name === 'Unbekannt') this.filter = { ...this.filter, contractId: null }
+      else this.filter = { ...this.filter, contractId: this.constracts.find(e => e.label === cat_name)?.id }
+      key = 'contractId'
+    }
+    if (series_name === 'Publikationsart') {
+      if (cat_name === 'Unbekannt') this.filter = { ...this.filter, pubTypeId: null }
+      else this.filter = { ...this.filter, pubTypeId: this.pub_types.find(e => e.label === cat_name)?.id }
+      key = 'pubTypeId'
+    }
+    if (series_name === 'Verlag') {
+      if (cat_name === 'Unbekannt') this.filter = { ...this.filter, publisherId: null }
+      else this.filter = { ...this.filter, publisherId: this.publisher.find(e => e.label === cat_name)?.id }
+      key = 'publisherId'
+    }
+    this.filterChips.push({ text: series_name + ': ' + cat_name, key })
     this.loadData(this.costs);
   }
   applyAntiFilter(series_name: string, cat_name: string) {
     if (series_name === 'Art' && cat_name === 'sonstige') return;
-    this.filterText += series_name+': !'+cat_name +' ';
+    if (series_name === 'Gesperrt') return;
+    let key = undefined;
+    let id = undefined;
     if (series_name === 'Art') {
-      if (cat_name === 'corresponding') this.filter = {...this.filter, corresponding: false}
-    } 
-    if (series_name === 'Institut') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, notInstituteId: null}
-      else this.filter = {...this.filter,notInstituteId: this.institutes.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'OA-Kategorie') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, notOaCatId: null}
-      else this.filter = {...this.filter,notOaCatId: this.oa_cats.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'Vertrag') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, notContractId: null}
-      else this.filter = {...this.filter,notContractId: this.constracts.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'Publikationsart') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, notPubTypeId: null}
-      else this.filter = {...this.filter,notPubTypeId: this.pub_types.find(e => e.label === cat_name)?.id}
-    } 
-    if (series_name === 'Verlag') {
-      if (cat_name === 'Unbekannt') this.filter = {...this.filter, notPublisherId: null}
-      else this.filter = {...this.filter,notPublisherId: this.publisher.find(e => e.label === cat_name)?.id}
+      if (cat_name === 'corresponding') this.filter = { ...this.filter, corresponding: false }
+      key = 'corresponding'
     }
+    if (series_name === 'Institut') {
+      let notInstituteId = this.filter?.notInstituteId ? this.filter.notInstituteId : [];
+      if (cat_name === 'Unbekannt') id = null;
+      else id = this.institutes.find(e => e.label === cat_name)?.id
+      notInstituteId.push(id)
+      this.filter.notInstituteId = notInstituteId;
+      key = 'notInstituteId'
+    }
+    if (series_name === 'OA-Kategorie') {
+      let notOaCatId = this.filter?.notOaCatId ? this.filter.notOaCatId : [];
+      if (cat_name === 'Unbekannt') id = null
+      else id = this.oa_cats.find(e => e.label === cat_name)?.id
+      notOaCatId.push(id)
+      this.filter.notOaCatId = notOaCatId;
+      key = 'notOaCatId'
+    }
+    if (series_name === 'Vertrag') {
+      let notContractId = this.filter?.notContractId ? this.filter.notContractId : [];
+      if (cat_name === 'Unbekannt') id=null
+      else id = this.constracts.find(e => e.label === cat_name)?.id;
+      notContractId.push(id)
+      this.filter.notContractId = notContractId;
+      key = 'notContractId'
+    }
+    if (series_name === 'Publikationsart') {
+      let notPubTypeId = this.filter?.notPubTypeId ? this.filter.notPubTypeId : [];
+      if (cat_name === 'Unbekannt') id = null
+      else id=this.pub_types.find(e => e.label === cat_name)?.id
+      notPubTypeId.push(id)
+      this.filter.notPubTypeId = notPubTypeId;
+      key = 'notPubTypeId'
+    }
+    if (series_name === 'Verlag') {
+      let notPublisherId = this.filter?.notPublisherId ? this.filter.notPublisherId : [];
+      if (cat_name === 'Unbekannt') id=null
+      else id=this.publisher.find(e => e.label === cat_name)?.id
+      notPublisherId.push(id)
+      this.filter.notPublisherId = notPublisherId;
+      key = 'notPublisherId'
+    }
+    this.filterChips.push({ text: series_name + ': !' + cat_name, key, id })
     this.loadData(this.costs);
   }
 
+  hidden() {
+    return !this.filter || Object.keys(this.filter).length === 0;
+  }
+
   resetFilter() {
-    this.filterText = '';
+    this.filterChips = [];
     this.filter = {};
+    this.loadData(this.costs);
+  }
+
+  removeFilter(elem: { text: string, key: string, id?: number }) {
+    if (elem.key.includes("not")) {
+      this.filter[elem.key] = this.filter[elem.key].filter(e => e !== elem.id)
+      this.filterChips = this.filterChips.filter(e => !(e.key === elem.key && e.id === elem.id))
+    }
+    else {
+      delete this.filter[elem.key]
+      this.filterChips = this.filterChips.filter(e => e.key !== elem.key)
+    }
     this.loadData(this.costs);
   }
 }
