@@ -26,44 +26,89 @@ export class CombineDialogComponent<T> implements OnInit {
   action(pos: number) {
     let res = this.ents[pos];
     if (this.data.aliases) {
-      //open alias dialog
-      let aliases = [];
-      for (let i = 0; i < this.ents.length; i++) {
-        if (i === pos) continue;
-        aliases.push(this.ents[i]['label']?.toLowerCase())
-      }
-
-      let dialogRef = this.dialog.open(AliasFormComponent, {
-        width: '400px',
-        maxHeight: '800px',
-        data: {
-          aliases
-        },
-        disableClose: true
-      });
-      dialogRef.afterClosed().subscribe(result => {
-        if (result) {
-          res = { ...res, aliases:result }
-          this.dialogRef.close(res);
+      if (!this.data.author) {
+        //open alias dialog
+        let aliases = [];
+        for (let i = 0; i < this.ents.length; i++) {
+          if (i === pos) continue;
+          aliases.push(this.ents[i]['label']?.toLowerCase())
         }
-        else this.dialogRef.close(res);
-      });
-    } else {
-      this.dialogRef.close(res);
+
+        let dialogRef = this.dialog.open(AliasFormComponent, {
+          width: '400px',
+          maxHeight: '800px',
+          data: {
+            aliases
+          },
+          disableClose: true
+        });
+        dialogRef.afterClosed().subscribe(result => {
+          if (result) {
+            res = { ...res, aliases: result }
+            this.dialogRef.close(res);
+          }
+          else this.dialogRef.close(res);
+        });
+      } else {//author
+        //open alias dialog first_name
+        let aliases = [];
+        for (let i = 0; i < this.ents.length; i++) {
+          if (i === pos || res['first_name'].toLowerCase().includes(this.ents[i]['first_name']?.toLowerCase())) continue;
+          aliases.push(this.ents[i]['first_name']?.toLowerCase())
+        }
+
+        let dialogRef = this.dialog.open(AliasFormComponent, {
+          width: '400px',
+          maxHeight: '800px',
+          data: {
+            aliases
+          },
+          disableClose: true
+        });
+        dialogRef.afterClosed().subscribe(result => {
+          if (result && result.length > 0) {
+            res = { ...res, aliases_first_name: result }
+          }
+          //open alias dialog last_name
+          let aliases = [];
+          for (let i = 0; i < this.ents.length; i++) {
+            if (i === pos || res['last_name'].toLowerCase().includes(this.ents[i]['last_name']?.toLowerCase())) continue;
+            aliases.push(this.ents[i]['last_name']?.toLowerCase())
+          }
+
+          let dialogRef = this.dialog.open(AliasFormComponent, {
+            width: '400px',
+            maxHeight: '800px',
+            data: {
+              aliases
+            },
+            disableClose: true
+          });
+          dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+              res = { ...res, aliases_last_name: result }
+            }
+            this.dialogRef.close(res);
+          });
+        }
+      );
     }
+  } else {
+  this.dialogRef.close(res);
+}
   }
 
-  abort() {
-    this.dialogRef.close(null);
-  }
+abort() {
+  this.dialogRef.close(null);
+}
 
-  getStyle(att: string) {
-    let style = "background:red;";
-    let value = "";
-    for (let ent of this.ents) {
-      if (!value) value = ent[att];
-      else if (ent[att] != value) return style;
-    }
-    return "";
+getStyle(att: string) {
+  let style = "background:red;";
+  let value = "";
+  for (let ent of this.ents) {
+    if (!value) value = ent[att];
+    else if (ent[att] != value) return style;
   }
+  return "";
+}
 }
