@@ -38,13 +38,18 @@ export class FilterViewComponent implements OnInit {
     { key: 'id', label: 'ID', type: 'number' },
     { key: 'title', label: 'Titel' },
     { key: 'doi', label: 'DOI' },
+    { key: 'other_ids', label: 'Weitere Identifikatoren' },
     { key: 'authors', label: 'Autor*innen' },
-    { key: 'inst_authors', label: 'Autor*innen der Institution' },
+    { key: 'inst_authors', label: 'Personen der Institution' },
+    { key: 'author_id', label: 'ID einer Person der Institution' },
+    { key: 'author_id_corr', label: 'ID einer Person der Institution (corr.)' },
     { key: 'institute', label: 'Institute' },
     { key: 'pub_date', label: 'Publikationsdatum', type: 'date' },
     { key: 'pub_date_accepted', label: 'Datum der Akzeptanz', type: 'date' },
     { key: 'greater_entity', label: 'Größere Einheit' },
+    { key: 'greater_entity_id', label: 'ID einer größeren Einheit' },
     { key: 'oa_category', label: 'OA-Kategorie' },
+    { key: 'oa_category_id', label: 'ID einer OA-Kategorie' },
     { key: 'dataSource', label: 'Datenquelle' },
     { key: 'language', label: 'Sprache' },
     { key: 'secound_pub', label: 'Zweitveröffentlichung' },
@@ -52,9 +57,13 @@ export class FilterViewComponent implements OnInit {
     { key: 'locked', label: 'Gesperrt', type: 'boolean' },
     { key: 'status', label: 'Status', type: 'number' },
     { key: 'pub_type', label: 'Publikationstyp' },
+    { key: 'pub_type_id', label: 'ID eines Publikationstyps' },
     { key: 'publisher', label: 'Verlag' },
+    { key: 'publisher_id', label: 'ID eines Verlags' },
     { key: 'contract', label: 'Vertrag' },
+    { key: 'contract_id', label: 'ID eines Vertrags' },
     { key: 'funder', label: 'Förderer' },
+    { key: 'funder_id', label: 'ID eines Förderer' },
     { key: 'edit_date', label: 'Letzte Bearbeitung', type: 'date' },
     { key: 'import_date', label: 'Importdatum', type: 'date' },
   ]
@@ -71,7 +80,12 @@ export class FilterViewComponent implements OnInit {
       next: data => {
         if (data['editors']) this.keys.push({ key: 'editors', label: 'Herausgeber*innen' })
         if (data['abstract']) this.keys.push({ key: 'abstract', label: 'Abstract' })
-        if (data['citation']) this.keys.push({ key: 'citation', label: 'Zitationsangabe' })
+        if (data['citation']) {
+          this.keys.push({ key: 'volume', label: 'Volume', type: 'number' })
+          this.keys.push({ key: 'issue', label: 'Issue', type: 'number' })
+          this.keys.push({ key: 'first_page', label: 'Erste Seite', type: 'number' })
+          this.keys.push({ key: 'last_page', label: 'Letzte Seite', type: 'number' })
+        }
         if (data['page_count']) this.keys.push({ key: 'page_count', label: 'Seitenzahl', type: 'number' })
         if (data['peer_reviewed']) this.keys.push({ key: 'peer_reviewed', label: 'Peer-Reviewed', type: 'boolean' })
         if (data['pub_date_print']) this.keys.push({ key: 'pub_date_print', label: 'Publikationsdatum (print)', type: 'date' })
